Move config to subdirectory
