+++ server-new/server.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"grocify-server/config"
 	"grocify-server/graph"
 	"log"
 	"net/http"
@@ -10,14 +11,18 @@ import (
 )
 
 func main() {
-	config := LoadConfig()
+	cfg := config.LoadConfig()
 
-	server := handler.NewDefaultServer(graph.NewExecutableSchema(graph.Config{Resolvers: &graph.Resolver{}}))
+	server := handler.NewDefaultServer(
+		graph.NewExecutableSchema(
+			graph.Config{ Resolvers: &graph.Resolver{} },
+		),
+	)
 
 	http.Handle("/", playground.Handler("GraphQL playground", "/query"))
 	http.Handle("/query", server)
 
 	log.Printf("Server up and running")
-	log.Printf("GraphQL playground available at http://%s:%s/", config.Host, config.Port)
-	log.Fatal(http.ListenAndServe(config.Host + ":" + config.Port, nil))
+	log.Printf("GraphQL playground available at http://%s:%s/", cfg.Host, cfg.Port)
+	log.Fatal(http.ListenAndServe(cfg.Host + ":" + cfg.Port, nil))
 }
